+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat: Error responses now include inner exception messages when `DetailedExceptionMessages` is enabled. (#468)
 - feat(c-admin-table): Clicking a row takes you to the details page (#465)
 - feat(c-admin-table): Always show button for details page (#465)
+- fix(c-select): Not respecting disabled/readonly state in all cases when opening menu (#473)
 
 - refactor: `CoalesceOptions.DetailedEntityFrameworkExceptionMessages` has been renamed to `CoalesceOptions.DetailedEFMigrationExceptionMessages`
 - fix: The "Max _N_ items retrieved" message in c-select now accounts for list calls that don't provide a count, e.g. by passing `noCount=true`.